ui: window/Settings: fix labelWidth for saved user name
Somewhere during rendering of the window, extjs wants to add 5px to
the labelwidth for the surrounding container. Having a string here
results in the container having a width of 1505px instead of 155px and
the value of the field is far off to the right and not visible.

Changing to a number fixes that.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>