ui: rework global searchfield, drop cruft, improve readability
Single letter variable names really do not help understanding what's
going on, as do overly general names like fields for the split up
words we actually search+filter for.
Using a switch block as map is often also not ideal (way more syntax
noise and style hacks like break on the same line to keep it
compact), rather just use an actual object map.

Some of the improvements where not possible when this was
implemented, as then we ensured < es5 compat for IE 10 support.

With for-of and nullish-chaining a few things to get nicer to
express.

While at it also fix comment text width making them less look like
haikus and improve match calculation comment to avoid implying that
match is either 0, 1 or 2, it can be higher too (if multiple columns
match).

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>