ui: ceph install wizard: fix config exists note position
During the Ceph installation, in the configuration step, a note is shown
if there already is a Ceph configuration present. This notification
should be in the center of the wizard but is currently barely visible.

ExtJS is having trouble calculating the position and the result is that
the note is placed almost out of the visible area.

Setting a fixed height helps ExtJS calculate the vertical position.
Changing from the 'beforeshow' listener to the 'afterrender' one helps
for the horizontal position.

Signed-off-by: Aaron Lauterer <a.lauterer@proxmox.com>