ui: add version to extjs script urls
So that we circumvent browsers caching 6.0 extjs js/css
this should (at least for new users upgrading) fix the browser caching
issue for extjs (we had some now in the forums)

I did it this way since we do not often change version of the extjs
package (since its a big task everytime anyway), but if wanted i can
do it differently. e.g.:
    * hardcode it in the perl code
    * generate it during the build (also for the control file)

But this is fine for ExtJS as we rarely update that and especially for
major releases we would need to adapt stuff anyway

Also bump the versioned dependency on extjs to 7.0.0 in the Debian
control file.

Signed-off-by: Dominik Csapak <d.csapak@proxmox.com>
Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>