ui: restore: improve warning for restoring container with same ID
It's not clear to users that the "VM data" includes mount point
volumes including those that are not marked for backup. This is
different from VM restore, where volumes attached at drives not
present in the backup will be kept around as unused volumes.

Several (supposedly newer) users got tripped up by this over the
years, the latest report being [0]. The long term plan is to make the
restore dialog more flexible to be able to select actions for disks
individually, but that will take a bit of time. In the mean time, make
the warning more explicit.

[0]: https://forum.proxmox.com/threads/111760/#post-482045

Signed-off-by: Fabian Ebner <f.ebner@proxmox.com>